refactor processing pipeline
